Add pipeline scripts

+++ pipeline/nfo_to_json.py
@@ -8,23 +8,37 @@ expected to be well-formed, so little-to-no error-checking is done.
 """
 import argparse
 import json
+import os
 import sys
 import xml.etree.ElementTree as ET
+from pathlib import Path
 
 def get_node_text(node):
     return node.text
 
+def get_node_int(node):
+    return int(node.text)
 
-tvshow_spec = {
-    "title": {"default": "Untitled", "multiple": False, "transform": get_node_text},
-    "plot": {"default": "No plot information", "multiple": False, "transform": get_node_text},
-    "year": {"default": 0, "multiple": False, "transform": lambda node: int(node.text)},
-    "genre": {"default": [], "multiple": True, "transform": get_node_text},
+
+specs = {
+    "tvshow": {
+        "title": {"default": "Untitled", "multiple": False, "transform": get_node_text},
+        "plot": {"default": "No plot information", "multiple": False, "transform": get_node_text},
+        "year": {"default": 0, "multiple": False, "transform": get_node_int},
+        "genre": {"default": [], "multiple": True, "transform": get_node_text},
+    },
+    "episodedetails": {
+        "title": {"default": "Untitled", "multiple": False, "transform": get_node_text},
+        "showtitle": {"default": "Untitled", "multiple": False, "transform": get_node_text},
+        "season": {"default": 0, "multiple": False, "transform": get_node_int},
+        "episode": {"default": 0, "multiple": False, "transform": get_node_int},
+        "plot": {"default": "No plot information", "multiple": False, "transform": get_node_text},
+    },
 }
 
 def parse_args():
     parser = argparse.ArgumentParser()
-    parser.add_argument("input", help="Input .nfo file")
+    parser.add_argument("input", help="Input .nfo file or directory containing .nfo files")
     parser.add_argument("-o", "--output", help="Output file (defaults to stdout)")
 
     return parser.parse_args()
@@ -32,49 +46,68 @@ def parse_args():
 
 # Ubuntu 18.04 doesn't use Python 3.7 by default. For now we'll return a dict
 # rather than using a dataclass.
-def parse_tvshow(xml: ET.Element):
-    tvshow = default_tvshow()
+def parse_nfo(xml: ET.Element, spec):
+    info = default_info(spec)
     for child in xml:
-        tag_spec = tvshow_spec.get(child.tag)
+        tag_spec = spec.get(child.tag)
         if tag_spec:
             tag_value = tag_spec["transform"](child)
             if tag_spec["multiple"]:
-                tvshow.setdefault(child.tag, []).append(tag_value)
+                info.setdefault(child.tag, []).append(tag_value)
             else:
-                tvshow[child.tag] = tag_value
+                info[child.tag] = tag_value
         else:
             # Print on stderr so that we don't interfere with normal shell redirects
             print(f"Ignoring tag {child.tag}", file=sys.stderr)
 
-    return tvshow
+    return info
 
 
 # This is a function since we always want a copy of the default
-def default_tvshow():
-    tvshow = {
+def default_info(spec):
+    info = {
         tag_name: tag_spec["default"]
-        for tag_name, tag_spec in tvshow_spec.items()
+        for tag_name, tag_spec in spec.items()
     }
 
-    return tvshow
+    return info
+
+
+def nfo_to_json(input_file, output_file):
+    root = ET.fromstring(input_file.read())
+
+    info = parse_nfo(root, specs[root.tag])
+    info_json = json.dumps(info, indent=4)
+
+    output_file.write(info_json)
+    output_file.write("\n")
 
 
 if __name__ == "__main__":
     args = parse_args()
 
-    tree = ET.parse(args.input)
-    root = tree.getroot()
+    path = Path(args.input)
 
-    assert root.tag == 'tvshow'
+    assert path.exists(), f"Path {path} does not exist"
 
-    tvshow = parse_tvshow(root)
-    tvshow_json = json.dumps(tvshow, indent=4)
+    if path.is_file():
+        with open(path) as input_file:
+            if args.output:
+                output_file = open(args.output, "w")
+                nfo_to_json(input_file, output_file)
+                output_file.close()
+            else:
+                output_file = sys.stdout
+                nfo_to_json(input_file, output_file)
 
-    if args.output:
-        output_file = open(args.output, 'w')
     else:
-        output_file = sys.stdout
+        for root, dirs, files in os.walk(path):
+            for name in files:
+                if Path(name).suffix != ".nfo":
+                    continue
+                nfo_path = Path(root) / name
+                json_path = nfo_path.with_suffix(".json")
 
-    output_file.write(tvshow_json)
-    output_file.write("\n")
-    output_file.close()
+                with open(nfo_path) as input_file:
+                    with open(json_path, "w") as output_file:
+                        nfo_to_json(input_file, output_file)
